backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets>=13,<16
python-dotenv==1.0.1
supabase>=2.28.0
pydantic-settings==2.1.0
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.4.0
groq==0.4.2
twilio==9.0.0
python-multipart==0.0.9
